requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
mypy-protobuf==3.1.0
protobuf==4.0.0rc2
psycopg2==2.9.3
requests==2.27.1
six==1.16.0
types-futures==3.3.7
types-protobuf==3.19.5
urllib3==1.26.8
websocket-client==1.2.3
